requirements.txt
# OpenAI Embeddings + Classifier Project
# Python 3.11+

openai>=1.0.0
pandas>=2.0.0
openpyxl>=3.1.0
scikit-learn>=1.3.0
joblib>=1.3.0
tenacity>=8.2.0
python-dotenv>=1.0.0
numpy>=1.24.0
lightgbm>=4.0.0
